refactor(enum): enums to folder
