--- src/app/shared/models/sign-up-steps.ts
@@ -1,6 +1,0 @@
-export enum SignUpSteps {
-  SensorCode = 'SensorCode',
-  UserType = 'UserType',
-  Register = 'Register',
-  FillInformation = 'FillInformation'
-}
